Cleaned up applicants models
